fix v1 integration tests
